requirements.txt
Flask
Flask-Alembic
Flask-JWT-Extended==3.24.1
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask-Login
openpyxl
Jinja2
SQLAlchemy
Werkzeug
flask_script
PyMysqlDB
mysqlclient
apscheduler
requests
opencv-python
pika
python-daemon
python-dateutil

